move core library into a dir
