+++ src/nanoarrow/common/array_stream.c
