+++ src/nanoarrow/common/inline_buffer.h
@@ -22,7 +22,7 @@
 #include <stdint.h>
 #include <string.h>
 
-#include "nanoarrow/inline_types.h"
+#include "nanoarrow/common/inline_types.h"
 
 #ifdef __cplusplus
 extern "C" {
