+++ src/nanoarrow/common/schema.c
